requirement.txt
alembic==1.3.1
autopep8==1.4.4
blinker==1.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Dance==3.0.0
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
oauthlib==3.1.0
pkg-resources==0.0.0
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pycparser==2.19
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
SQLAlchemy==1.3.11
SQLAlchemy-Utils==0.35.0
urllib3==1.25.7
URLObject==2.4.3
Werkzeug==0.16.0
